+++ src/app/shared/components/add-to-list/add-to-list.component.ts
@@ -1,10 +1,12 @@
 import { Component, Input, OnInit } from '@angular/core';
+import { MatDialog, MatDialogRef } from '@angular/material/dialog';
 import { Store } from '@ngrx/store';
 import { Observable } from 'rxjs';
 import { AppState } from 'src/app/app.store';
 import { IListDetails } from 'src/app/lists/interfaces/list-details-response.interface';
 import { addMovieToList, deleteMovieFromList } from 'src/app/lists/store/lists.actions';
 import { selectAllLists } from 'src/app/lists/store/lists.selector';
+import { ConfirmationDialogComponent } from '../confirmation-dialog/confirmation-dialog.component';
 
 @Component({
   selector: 'app-add-to-list',
@@ -14,11 +16,21 @@ export class AddToListComponent implements OnInit {
   @Input() movieId!: number;
   lists$!: Observable<IListDetails[]>;
   dropdownHidden = true;
+  confirmDialog!: MatDialogRef<ConfirmationDialogComponent>;
 
-  constructor(private store: Store<AppState>) {}
+  constructor(private store: Store<AppState>, private dialog: MatDialog) {}
 
   ngOnInit(): void {
     this.lists$ = this.store.select(selectAllLists);
+  }
+
+  openConfirmationDialog(question: string): Observable<boolean> {
+    this.confirmDialog = this.dialog.open(ConfirmationDialogComponent, {
+      disableClose: true,
+    });
+    this.confirmDialog.componentInstance.dialogMessage =
+      question;
+    return this.confirmDialog.afterClosed();
   }
 
   addMovieToList(listId: string) {
@@ -28,9 +40,13 @@ export class AddToListComponent implements OnInit {
   }
 
   removeMovieFromList(listId: string){
-    this.store.dispatch(
-      deleteMovieFromList({movieId : this.movieId, listId: parseInt(listId)})
-    )
+    this.openConfirmationDialog("Are you sure you want to remove this movie from the list?").subscribe(result => {
+      if(result){
+        this.store.dispatch(
+          deleteMovieFromList({movieId : this.movieId, listId: parseInt(listId)})
+        )
+      }
+    })
   }
 
   isMovieInList(list: IListDetails): boolean {
